requirements.txt
matplotlib==3.5.1
numpy==1.22.4
opencv_python==4.4.0.44
pandas==1.2.4
Pillow==9.5.0
pynput==1.7.6
pywin32==227
PyYAML==6.0
rawpy==0.17.1
scikit_image==0.16.2
scikit_learn==1.0.2
scipy==1.7.1
tensorboardX==2.5
tensorboardX==2.6
torch==1.8.1+cu101
torchgeometry==0.1.2
torchvision==0.9.1+cu101
tqdm==4.62.3
